==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
import pandas as pd
rides = pd.read_csv('../data/transformed/validated_rides_2024.csv', index_col=0)
rides.head(10)

rides['pickup_datetime'] = pd.to_datetime(rides['pickup_datetime'], format='%m/%d/%Y %I:%M:%S %p')
rides['pickup_hour'] = rides['pickup_datetime'].dt.round('h')


rides.head()

rides = rides[rides['pickup_latitude'].isnull()==False]

import pandas as pd
import geopandas as gpd
from shapely.wkt import loads

chicago_zones = pd.read_csv('../data/chicago_geo_data.csv')
chicago_zones["geometry"] = chicago_zones["the_geom"].apply(loads)
chicago_zones = gpd.GeoDataFrame(chicago_zones, geometry="geometry", crs="EPSG:4326")

# Verificar que las geometrías sean correctas
print(chicago_zones.head())


from shapely.geometry import Point

# Convertir el DataFrame de taxis en un GeoDataFrame
geometry = [Point(xy) for xy in zip(rides['pickup_longitude'], rides['pickup_latitude'])]
rides_gdf = gpd.GeoDataFrame(rides, geometry=geometry, crs="EPSG:4326")

# Hacer el join espacial para asignar cada taxi a un barrio
rides_with_zones = gpd.sjoin(rides_gdf, chicago_zones[['PRI_NEIGH', 'geometry']], 
                             how="left", predicate="within")

# Mantener solo las columnas originales de rides + el barrio asignado
rides_with_zones = rides_with_zones[rides.columns.tolist() + ['PRI_NEIGH']]

# Rellenar valores nulos con "Outside Chicago"
rides_with_zones['neighborhood_name'] = rides_with_zones['PRI_NEIGH'].fillna("Outside Chicago")
rides_with_zones.drop(columns='PRI_NEIGH', inplace=True)
# Ver los primeros resultados
print(rides_with_zones.head())


rides_with_zones

agg_rides = rides_with_zones.groupby(['pickup_hour', 'neighborhood_name']).size().reset_index()
agg_rides.rename(columns={0:'rides'}, inplace=True)
agg_rides.sort_values(by='rides')

from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    
    location_ids = agg_rides['neighborhood_name'].unique()
    full_range = pd.date_range(
        agg_rides['pickup_hour'].min(), agg_rides['pickup_hour'].max(), freq='H')
    output = pd.DataFrame()
    for location_id in tqdm(location_ids):

        # keep only rides for this 'location_id'
        agg_rides_i = agg_rides.loc[agg_rides.neighborhood_name == location_id, ['pickup_hour', 'rides']]
            
        # quick way to add missing dates with 0 in a Series
        # taken from https://stackoverflow.com/a/19324591
        agg_rides_i.set_index('pickup_hour', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)
        
        # add back `location_id` columns
        agg_rides_i['neighborhood_name'] = location_id

        output = pd.concat([output, agg_rides_i])
    
    # move the purchase_day from the index to a dataframe column
    output = output.reset_index().rename(columns={'index': 'pickup_hour'})
    
    return output

agg_rides_all_slots = add_missing_slots(agg_rides)
agg_rides_all_slots

agg_rides_all_slots.neighborhood_name.unique()

from typing import Optional, List
import pandas as pd
import plotly.express as px

def plot_rides(
    rides: pd.DataFrame,
    locations: Optional[List[int]] = None,
    start_date: Optional[str] = None,
    end_date: Optional[str] = None
):
    """
    Plot time-series data with optional location and date filtering.

    :param rides: DataFrame with ride data.
    :param locations: List of neighborhoods to filter (Optional).
    :param start_date: Start date as 'YYYY-MM-DD' (Optional).
    :param end_date: End date as 'YYYY-MM-DD' (Optional).
    """
    rides_to_plot = rides.copy()
    
    # Filtrar por ubicación si se especifica
    if locations:
        rides_to_plot = rides_to_plot[rides_to_plot.neighborhood_name.isin(locations)]
    
    # Filtrar por rango de fechas si se especifica
    if start_date:
        rides_to_plot = rides_to_plot[rides_to_plot["pickup_hour"] >= start_date]
    if end_date:
        rides_to_plot = rides_to_plot[rides_to_plot["pickup_hour"] <= end_date]

    # Graficar
    fig = px.line(
        rides_to_plot,
        x="pickup_hour",
        y="rides",
        color="neighborhood_name",
        template="none",
    )

    fig.show()


plot_rides(agg_rides_all_slots, locations=["Gold Coast", "Loop", "O'Hare"], start_date="2024-01-01", end_date="2024-02-01")


agg_rides_all_slots.to_csv("../data/transformed/ts_data_2024.csv")
