==> notebooks/11_backfill_feature_store.ipynb <==
HOPSWORKS_PROJECT_NAME = 'chicago_taxi_demand'

import os
from dotenv import load_dotenv
from src.paths import PARENT_DIR

load_dotenv(PARENT_DIR / '.env')
HOPSWORKS_API_KEY = os.environ['HOPSWORKS_API_KEY']

from datetime import datetime
import pandas as pd
from src.data import load_raw_data

rides = load_raw_data()
rides

from src.data import merge_geo_and_ts_data
rides = merge_geo_and_ts_data(rides)
rides

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

# string to datetime
ts_data['pickup_hour'] = pd.to_datetime(ts_data['pickup_hour'], utc=True)

# add column with Unix epoch milliseconds
ts_data['pickup_ts'] = ts_data['pickup_hour'].astype('int64') // 10**6

ts_data.info()

import hopsworks

project = hopsworks.login(
    project=HOPSWORKS_PROJECT_NAME,
    api_key_value=HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()

feature_store

FEATURE_GROUP_NAME = 'time_series_hourly_feature_group'
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ['pickup_location', 'pickup_hour'],
    event_time='pickup_hour',
)

#This cell is because an error when uploading data

from confluent_kafka import Producer, Consumer, TopicPartition
import hsfs.core.kafka_engine

# Forzar la asignación de Producer y Consumer
hsfs.core.kafka_engine.Producer = Producer  
hsfs.core.kafka_engine.Consumer = Consumer  
hsfs.core.kafka_engine.TopicPartition = TopicPartition

feature_group.insert(ts_data,write_options={"wait_for_job":False})
