==> notebooks/12_feature_pipeline.ipynb <==
import src.config as config

from datetime import datetime, timedelta

import pandas as pd

current_date = pd.Timestamp.now(tz="UTC").floor('H')
print(f'{current_date=}')

# we fetch raw data for the last 28 days, to add redundancy to our data pipeline
fetch_data_to = current_date
fetch_data_from = max(current_date - timedelta(days=28), pd.Timestamp("2024-01-01", tz="UTC"))
print(fetch_data_from, fetch_data_to)

from src.data import load_raw_data
from src.data import merge_geo_and_ts_data


def fetch_batch_raw_data(from_date: datetime, to_date: datetime) -> pd.DataFrame:
    """
    Simula producción de datos pero ajustándose al rango disponible (desde enero 2024).
    """
    # Intentamos retroceder 52 semanas, pero respetamos el límite mínimo de enero 2024
    from_date_ = max(from_date - timedelta(days=7*52), pd.Timestamp("2024-01-01", tz="UTC"))
    to_date_ = max(to_date - timedelta(days=7*52), pd.Timestamp("2024-01-01", tz="UTC"))
    
    print(f'{from_date=}, {to_date_=}')

    # Cargar solo los datos dentro del rango disponible
    rides = load_raw_data()
    rides = merge_geo_and_ts_data(rides)

    # Solo ajustamos la fecha si no estamos en el límite inferior
    if from_date_ > pd.Timestamp("2024-01-01", tz="UTC"):
        rides['pickup_datetime'] += timedelta(days=7*52)

    rides.sort_values(by=['pickup_location', 'pickup_datetime'], inplace=True)

    return rides



rides = fetch_batch_raw_data(from_date=fetch_data_from, to_date=fetch_data_to)

rides

from src.data import transform_raw_data_into_ts_data
ts_data = transform_raw_data_into_ts_data(rides)

# string to datetime
ts_data['pickup_hour'] = pd.to_datetime(ts_data['pickup_hour'], utc=True)

# add column with Unix epoch milliseconds
ts_data['pickup_ts'] = ts_data['pickup_hour'].astype('int64') // 10**6

import hopsworks

# connect to the project
project = hopsworks.login(
    project=config.HOPSWORKS_PROJECT_NAME,
    api_key_value=config.HOPSWORKS_API_KEY
)

# connect to the feature store
feature_store = project.get_feature_store()

# connect to the feature group
feature_group = feature_store.get_or_create_feature_group(
    name=config.FEATURE_GROUP_NAME,
    version=config.FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ['pickup_location', 'pickup_hour'],
    event_time='pickup_hour',
)

feature_group.insert(ts_data, write_options={"wait_for_job": True})
